Check SSH connection is to a known host
Go changed SSH behaviour, so by default it is necessary to validate the
host key.

This necessitated changing the ":ssh" to ":22" which is equivalent.
Without doing so, the `HostKeyCallback` from `knownhosts` doesn't
correctly pick up the entry in the `known_hosts` file. This is probably
because, otherwise, the host becomes `example.com:ssh` and no match is
found in the SSH known hosts.

If the host is not known, then we error out. This seems
sensible from a security point of view. If you need to connect to a
server, then simply SSH in once first, validating the server's identity
then.